backend/requirements.txt
astroid==2.2.5
autopep8==1.4.4
backports.csv==1.0.7
certifi==2019.3.9
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-cors-headers==2.5.3
django-filter==2.1.0
django-import-export==1.2.0
djangorestframework==3.9.3
et-xmlfile==1.0.1
gunicorn==19.9.0
isort==4.3.4
jdcal==1.4.1
lazy-object-proxy==1.3.1
Markdown==3.1
mccabe==0.6.1
odfpy==1.4.0
openpyxl==2.6.2
pep8==1.7.1
psycopg2==2.7.6.1
pycodestyle==2.5.0
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pytz==2019.1
PyYAML==5.1.1
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
typed-ast==1.3.5
wrapt==1.10.11
xlrd==1.2.0
xlwt==1.3.0
